+++ packages/zwave-js/src/Utils.ts
@@ -1,2 +1,2 @@
-export { extractFirmware } from "@zwave-js/core";
+export { extractFirmware, guessFirmwareFileFormat } from "@zwave-js/core";
 export type { Firmware, FirmwareFileFormat } from "@zwave-js/core";