+++ packages/zwave-js/src/lib/commandclass/MultiChannelAssociationCC.ts
@@ -27,6 +27,7 @@ import {
 	CommandClass,
 	commandClass,
 	CommandClassDeserializationOptions,
+	CommandClassOptions,
 	expectedCCResponse,
 	gotDeserializationOptions,
 	implementedVersion,
@@ -207,9 +208,10 @@ export class MultiChannelAssociationCCAPI extends CCAPI {
 				endpoint: this.endpoint.index,
 			},
 		);
-		const response = (await this.driver.sendCommand<
-			MultiChannelAssociationCCSupportedGroupingsReport
-		>(cc, this.commandOptions))!;
+		const response = (await this.driver.sendCommand<MultiChannelAssociationCCSupportedGroupingsReport>(
+			cc,
+			this.commandOptions,
+		))!;
 		return response.groupCount;
 	}
 
@@ -228,9 +230,10 @@ export class MultiChannelAssociationCCAPI extends CCAPI {
 			endpoint: this.endpoint.index,
 			groupId,
 		});
-		const response = (await this.driver.sendCommand<
-			MultiChannelAssociationCCReport
-		>(cc, this.commandOptions))!;
+		const response = (await this.driver.sendCommand<MultiChannelAssociationCCReport>(
+			cc,
+			this.commandOptions,
+		))!;
 		return {
 			maxNodes: response.maxNodes,
 			nodeIds: response.nodeIds,
@@ -304,6 +307,14 @@ export class MultiChannelAssociationCCAPI extends CCAPI {
 @implementedVersion(4)
 export class MultiChannelAssociationCC extends CommandClass {
 	declare ccCommand: MultiChannelAssociationCommand;
+
+	public constructor(driver: Driver, options: CommandClassOptions) {
+		super(driver, options);
+		// Make valueIDs internal
+		this.registerValue(getMaxNodesValueId(0).property, true);
+		this.registerValue(getNodeIdsValueId(0).property, true);
+		this.registerValue(getEndpointsValueId(0).property, true);
+	}
 
 	public determineRequiredCCInterviews(): readonly CommandClasses[] {
 		// MultiChannelAssociationCC must be interviewed after Z-Wave+ if that is supported