+++ packages/zwave-js/src/lib/commandclass/MultilevelSensorCC.ts
@@ -56,6 +56,16 @@ export interface MultilevelSensorValue {
 	scale: Scale;
 }
 
+/**
+ * @publicAPI
+ */
+export type MultilevelSensorValueMetadata = ValueMetadata & {
+	ccSpecific: {
+		sensorType: number;
+		scale: number;
+	};
+};
+
 // @noSetValueAPI This CC is read-only
 
 @API(CommandClasses["Multilevel Sensor"])
@@ -87,9 +97,10 @@ export class MultilevelSensorCCAPI extends CCAPI {
 			sensorType,
 			scale,
 		});
-		const response = (await this.driver.sendCommand<
-			MultilevelSensorCCReport
-		>(cc, this.commandOptions))!;
+		const response = (await this.driver.sendCommand<MultilevelSensorCCReport>(
+			cc,
+			this.commandOptions,
+		))!;
 
 		if (sensorType === undefined) {
 			// Overload #1: return the full response
@@ -114,9 +125,10 @@ export class MultilevelSensorCCAPI extends CCAPI {
 			nodeId: this.endpoint.nodeId,
 			endpoint: this.endpoint.index,
 		});
-		const response = (await this.driver.sendCommand<
-			MultilevelSensorCCSupportedSensorReport
-		>(cc, this.commandOptions))!;
+		const response = (await this.driver.sendCommand<MultilevelSensorCCSupportedSensorReport>(
+			cc,
+			this.commandOptions,
+		))!;
 		return response.supportedSensorTypes;
 	}
 
@@ -133,9 +145,10 @@ export class MultilevelSensorCCAPI extends CCAPI {
 			endpoint: this.endpoint.index,
 			sensorType,
 		});
-		const response = (await this.driver.sendCommand<
-			MultilevelSensorCCSupportedScaleReport
-		>(cc, this.commandOptions))!;
+		const response = (await this.driver.sendCommand<MultilevelSensorCCSupportedScaleReport>(
+			cc,
+			this.commandOptions,
+		))!;
 		return response.sensorSupportedScales;
 	}
 
@@ -403,6 +416,10 @@ export class MultilevelSensorCCReport extends MultilevelSensorCC {
 			...ValueMetadata.ReadOnlyNumber,
 			unit: this.scale.unit,
 			label: typeName,
+			ccSpecific: {
+				sensorType: this.type,
+				scale: this.scale.key,
+			},
 		});
 		this.getValueDB().setValue(valueId, this.value);
 		return true;