+++ packages/zwave-js/src/lib/commandclass/MultilevelSwitchCC.ts
@@ -31,6 +31,7 @@ import {
 	CommandClass,
 	commandClass,
 	CommandClassDeserializationOptions,
+	CommandClassOptions,
 	expectedCCResponse,
 	gotDeserializationOptions,
 	implementedVersion,
@@ -81,11 +82,28 @@ const switchTypeProperties = Object.keys(SwitchType)
 	.map((key) => switchTypeToActions(key))
 	.reduce<string[]>((acc, cur) => acc.concat(...cur), []);
 
+/**
+ * @publicAPI
+ */
+export type MultilevelSwitchLevelChangeMetadata = ValueMetadata & {
+	ccSpecific: {
+		switchType: SwitchType;
+	};
+};
+
 function getCurrentValueValueID(endpoint: number): ValueID {
 	return {
 		commandClass: CommandClasses["Multilevel Switch"],
 		endpoint,
 		property: "currentValue",
+	};
+}
+
+/** Returns the ValueID used to remember whether a node supports supervision on the start/stop level change commands*/
+function getSuperviseStartStopLevelChangeValueId(): ValueID {
+	return {
+		commandClass: CommandClasses["Multilevel Switch"],
+		property: "superviseStartStopLevelChange",
 	};
 }
 
@@ -115,9 +133,10 @@ export class MultilevelSwitchCCAPI extends CCAPI {
 			nodeId: this.endpoint.nodeId,
 			endpoint: this.endpoint.index,
 		});
-		const response = (await this.driver.sendCommand<
-			MultilevelSwitchCCReport
-		>(cc, this.commandOptions))!;
+		const response = (await this.driver.sendCommand<MultilevelSwitchCCReport>(
+			cc,
+			this.commandOptions,
+		))!;
 		return {
 			currentValue: response.currentValue,
 			targetValue: response.targetValue,
@@ -185,20 +204,33 @@ export class MultilevelSwitchCCAPI extends CCAPI {
 			...options,
 		});
 
-		// Try to supervise the command execution
-		const supervisionResult = await this.driver.trySendCommandSupervised(
-			cc,
-		);
+		const superviseValueId = getSuperviseStartStopLevelChangeValueId();
+		const node = this.endpoint.getNodeUnsafe()!;
+		// Assume supervision is supported until we know it is not
+		let mayUseSupervision = node.getValue(superviseValueId) !== false;
 
-		if (
-			supervisionResult &&
-			(supervisionResult.status === SupervisionStatus.Fail ||
-				supervisionResult.status === SupervisionStatus.NoSupport)
-		) {
-			throw new ZWaveError(
-				"startLevelChange failed",
-				ZWaveErrorCodes.SupervisionCC_CommandFailed,
+		if (mayUseSupervision) {
+			// Try to supervise the command execution
+			const supervisionResult = await this.driver.trySendCommandSupervised(
+				cc,
 			);
+
+			if (supervisionResult?.status === SupervisionStatus.Fail) {
+				throw new ZWaveError(
+					"startLevelChange failed",
+					ZWaveErrorCodes.SupervisionCC_CommandFailed,
+				);
+			} else if (
+				supervisionResult?.status === SupervisionStatus.NoSupport
+			) {
+				// Remember that we shouldn't use supervision for that
+				node.valueDB.setValue(superviseValueId, false);
+				mayUseSupervision = false;
+			}
+		}
+		// In order to support a fallback to no supervision, we must not use else-if here
+		if (!mayUseSupervision) {
+			await this.driver.sendCommand(cc);
 		}
 	}
 
@@ -213,20 +245,33 @@ export class MultilevelSwitchCCAPI extends CCAPI {
 			endpoint: this.endpoint.index,
 		});
 
-		// Try to supervise the command execution
-		const supervisionResult = await this.driver.trySendCommandSupervised(
-			cc,
-		);
+		const superviseValueId = getSuperviseStartStopLevelChangeValueId();
+		const node = this.endpoint.getNodeUnsafe()!;
+		// Assume supervision is supported until we know it is not
+		let mayUseSupervision = node.getValue(superviseValueId) !== false;
 
-		if (
-			supervisionResult &&
-			(supervisionResult.status === SupervisionStatus.Fail ||
-				supervisionResult.status === SupervisionStatus.NoSupport)
-		) {
-			throw new ZWaveError(
-				"stopLevelChange failed",
-				ZWaveErrorCodes.SupervisionCC_CommandFailed,
+		if (mayUseSupervision) {
+			// Try to supervise the command execution
+			const supervisionResult = await this.driver.trySendCommandSupervised(
+				cc,
 			);
+
+			if (supervisionResult?.status === SupervisionStatus.Fail) {
+				throw new ZWaveError(
+					"stopLevelChange failed",
+					ZWaveErrorCodes.SupervisionCC_CommandFailed,
+				);
+			} else if (
+				supervisionResult?.status === SupervisionStatus.NoSupport
+			) {
+				// Remember that we shouldn't use supervision for that
+				node.valueDB.setValue(superviseValueId, false);
+				mayUseSupervision = false;
+			}
+		}
+		// In order to support a fallback to no supervision, we must not use else-if here
+		if (!mayUseSupervision) {
+			await this.driver.sendCommand(cc);
 		}
 	}
 
@@ -240,9 +285,10 @@ export class MultilevelSwitchCCAPI extends CCAPI {
 			nodeId: this.endpoint.nodeId,
 			endpoint: this.endpoint.index,
 		});
-		const response = (await this.driver.sendCommand<
-			MultilevelSwitchCCSupportedReport
-		>(cc, this.commandOptions))!;
+		const response = (await this.driver.sendCommand<MultilevelSwitchCCSupportedReport>(
+			cc,
+			this.commandOptions,
+		))!;
 		return response.switchType;
 	}
 
@@ -306,6 +352,14 @@ export class MultilevelSwitchCCAPI extends CCAPI {
 @implementedVersion(4)
 export class MultilevelSwitchCC extends CommandClass {
 	declare ccCommand: MultilevelSwitchCommand;
+
+	public constructor(driver: Driver, options: CommandClassOptions) {
+		super(driver, options);
+		this.registerValue(
+			getSuperviseStartStopLevelChangeValueId().property,
+			true,
+		);
+	}
 
 	public async interview(complete: boolean = true): Promise<void> {
 		const node = this.getNode()!;
@@ -393,12 +447,14 @@ export class MultilevelSwitchCC extends CommandClass {
 			this.getValueDB().setMetadata(upValueId, {
 				...ValueMetadata.Boolean,
 				label: `Perform a level change (${up})`,
+				ccSpecific: { switchType },
 			});
 		}
 		if (!valueDb.hasMetadata(downValueId)) {
 			this.getValueDB().setMetadata(downValueId, {
 				...ValueMetadata.Boolean,
 				label: `Perform a level change (${down})`,
+				ccSpecific: { switchType },
 			});
 		}
 	}