+++ packages/zwave-js/src/lib/commandclass/SoundSwitchCC.ts
@@ -98,9 +98,10 @@ export class SoundSwitchCCAPI extends CCAPI {
 			nodeId: this.endpoint.nodeId,
 			endpoint: this.endpoint.index,
 		});
-		const response = (await this.driver.sendCommand<
-			SoundSwitchCCTonesNumberReport
-		>(cc, this.commandOptions))!;
+		const response = (await this.driver.sendCommand<SoundSwitchCCTonesNumberReport>(
+			cc,
+			this.commandOptions,
+		))!;
 		return response.toneCount;
 	}
 
@@ -116,9 +117,10 @@ export class SoundSwitchCCAPI extends CCAPI {
 			endpoint: this.endpoint.index,
 			toneId,
 		});
-		const response = (await this.driver.sendCommand<
-			SoundSwitchCCToneInfoReport
-		>(cc, this.commandOptions))!;
+		const response = (await this.driver.sendCommand<SoundSwitchCCToneInfoReport>(
+			cc,
+			this.commandOptions,
+		))!;
 		return pick(response, ["duration", "name"]);
 	}
 
@@ -154,9 +156,10 @@ export class SoundSwitchCCAPI extends CCAPI {
 			nodeId: this.endpoint.nodeId,
 			endpoint: this.endpoint.index,
 		});
-		const response = (await this.driver.sendCommand<
-			SoundSwitchCCConfigurationReport
-		>(cc, this.commandOptions))!;
+		const response = (await this.driver.sendCommand<SoundSwitchCCConfigurationReport>(
+			cc,
+			this.commandOptions,
+		))!;
 		return pick(response, ["defaultToneId", "defaultVolume"]);
 	}
 
@@ -214,9 +217,10 @@ export class SoundSwitchCCAPI extends CCAPI {
 			nodeId: this.endpoint.nodeId,
 			endpoint: this.endpoint.index,
 		});
-		const response = (await this.driver.sendCommand<
-			SoundSwitchCCTonePlayReport
-		>(cc, this.commandOptions))!;
+		const response = (await this.driver.sendCommand<SoundSwitchCCTonePlayReport>(
+			cc,
+			this.commandOptions,
+		))!;
 		return pick(response, ["toneId", "volume"]);
 	}
 
@@ -627,6 +631,9 @@ export class SoundSwitchCCTonePlayReport extends SoundSwitchCC {
 		super(driver, options);
 		validatePayload(this.payload.length >= 1);
 		this.toneId = this.payload[0];
+		if (this.toneId !== 0 && this.payload.length >= 2) {
+			this.volume = this.payload[1];
+		}
 
 		this.persistValues();
 	}